+++ modd/Cargo.toml
@@ -6,7 +6,10 @@ edition = "2021"
 license = "GPL-2.0-or-later"
 
 [dependencies]
+game = { workspace = true }
 log = { workspace = true }
 rand = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+strum = { workspace = true }
+thiserror = { workspace = true }
