+++ randomizer/src/lib.rs
@@ -7,23 +7,22 @@ use std::{
 };
 
 use game::{
-    world::{Area, Group as GroupId, LocationKey},
-    Item::{self, *},
+    world::LocationKey,
+    Item::{self},
 };
-use log::{debug, error, info};
-use modd::Settings;
+use log::{error, info};
+use modd::{Layout, Settings};
 use model::filler_item::{convert, FillerItem};
 use patch::Patcher;
 use path_absolutize::*;
 use rand::{rngs::StdRng, SeedableRng};
 use rom::Rom;
-use serde::{ser::SerializeMap, Serialize, Serializer};
+use serde::{Serialize, Serializer};
 
 use crate::{
     constants::VERSION,
     hints::{formatting::*, Hints},
     metrics::Metrics,
-    patch::msbf::MsbfKey,
     system::UserConfig,
 };
 
@@ -59,330 +58,9 @@ impl Error {
     }
 }
 
-/// A world layout for the patcher.
-#[derive(Clone, Debug, Default, Serialize)]
-pub struct Layout {
-    #[serde(rename = "Hyrule", serialize_with = "serialize_world")]
-    hyrule: Group,
-    #[serde(rename = "Lorule", serialize_with = "serialize_world")]
-    lorule: Group,
-    #[serde(rename = "Dungeons", serialize_with = "serialize_world")]
-    dungeons: Group,
-}
-
-impl Layout {
-    fn group(&self, id: GroupId) -> &Group {
-        match id {
-            GroupId::Hyrule => &self.hyrule,
-            GroupId::Lorule => &self.lorule,
-            GroupId::Dungeons => &self.dungeons,
-        }
-    }
-
-    fn group_mut(&mut self, id: GroupId) -> &mut Group {
-        match id {
-            GroupId::Hyrule => &mut self.hyrule,
-            GroupId::Lorule => &mut self.lorule,
-            GroupId::Dungeons => &mut self.dungeons,
-        }
-    }
-
-    fn get_area_mut(&mut self, area: Area) -> &mut BTreeMap<&'static str, Item> {
-        self.group_mut(area.group()).entry(area.name()).or_insert_with(Default::default)
-    }
-
-    fn get(&self, key: &LocationKey) -> Option<Item> {
-        let LocationKey { area, name } = key;
-        self.group(area.group()).get(area.name()).and_then(|region| region.get(name).copied())
-    }
-
-    #[allow(unused)]
-    fn find(&self, item: Item) -> Vec<&'static str> {
-        todo!()
-    }
-
-    /// This just highlights why we need to redo [`Layout`]
-    #[allow(unused)]
-    fn find_single(&self, find_item: Item) -> Option<(&'static str, &'static str)> {
-        for (region_name, region) in &self.hyrule {
-            for (loc_name, item) in region {
-                if find_item.eq(item) {
-                    return Some((region_name, loc_name));
-                }
-            }
-        }
-
-        for (region_name, region) in &self.lorule {
-            for (loc_name, item) in region {
-                if find_item.eq(item) {
-                    return Some((region_name, loc_name));
-                }
-            }
-        }
-
-        for (region_name, region) in &self.dungeons {
-            for (loc_name, item) in region {
-                if find_item.eq(item) {
-                    return Some((region_name, loc_name));
-                }
-            }
-        }
-
-        None
-    }
-
-    pub fn set(&mut self, location: LocationKey, item: Item) {
-        let LocationKey { area: node, name } = location;
-        self.get_area_mut(node).insert(name, item.normalize());
-        debug!(
-            "Placed {} in {}/{}",
-            item.normalize().as_ref(),
-            location.area.name(),
-            location.name
-        );
-    }
-}
-
-pub type Group = BTreeMap<&'static str, BTreeMap<&'static str, Item>>;
-
-fn serialize_world<S>(region: &Group, ser: S) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    struct Wrap<'a>(&'a BTreeMap<&'static str, Item>);
-
-    impl<'a> Serialize for Wrap<'a> {
-        fn serialize<S>(&self, ser: S) -> Result<S::Ok, S::Error>
-        where
-            S: Serializer,
-        {
-            let mut map = ser.serialize_map(Some(self.0.len()))?;
-            for (k, v) in self.0 {
-                map.serialize_entry(k, item_to_str(v))?;
-            }
-            map.end()
-        }
-    }
-
-    let mut map = ser.serialize_map(Some(region.len()))?;
-    for (k, v) in region {
-        map.serialize_entry(k, &Wrap(v))?;
-    }
-    map.end()
-}
-
-fn item_to_str(item: &Item) -> &'static str {
-    match item {
-        KeySmall => "Small Key",
-        KeyBoss => "Big Key",
-        Compass => "Compass",
-        HeartContainer => "Heart Container",
-        RupeeR => "Red Rupee",
-        RupeeG => "Green Rupee",
-        RupeeB => "Blue Rupee",
-        HeartPiece => "Piece of Heart",
-        ItemIceRod => "Ice Rod",
-        ItemIceRodLv2 => "Nice Ice Rod",
-        ItemSandRod => "Sand Rod",
-        ItemSandRodLv2 => "Nice Sand Rod",
-        ItemTornadeRod => "Tornado Rod",
-        ItemTornadeRodLv2 => "Nice Tornado Rod",
-        ItemBomb => "Bombs",
-        ItemBombLv2 => "Nice Bombs",
-        ItemFireRod => "Fire Rod",
-        ItemFireRodLv2 => "Nice Fire Rod",
-        ItemHookShot => "Hookshot",
-        ItemHookShotLv2 => "Nice Hookshot",
-        ItemBoomerang => "Boomerang",
-        ItemBoomerangLv2 => "Nice Boomerang",
-        ItemHammer => "Hammer",
-        ItemHammerLv2 => "Nice Hammer",
-        ItemBow => "Bow",
-        ItemBowLv2 => "Nice Bow",
-        ItemShield => "Shield",
-        ItemBottle => "Empty Bottle",
-        ItemStoneBeauty => "Smooth Gem",
-        ItemKandelaar => "Lamp",
-        ItemKandelaarLv2 => "Super Lamp",
-        ItemSwordLv1 => "Sword Upgrade",
-        ItemSwordLv2 => "Sword Upgrade",
-        ItemSwordLv3 => "Master Sword Lv2",
-        ItemSwordLv4 => "Master Sword Lv3",
-        ItemMizukaki => "Zora's Flippers",
-        RingRental => "Bracelet Upgrade",
-        RingHekiga => "Ravio's Bracelet",
-        ItemBell => "Bell",
-        RupeeGold => "Gold Rupee",
-        RupeeSilver => "Silver Rupee",
-        PowerGlove => "Strength Upgrade",
-        ItemInsectNet => "Net",
-        ItemInsectNetLv2 => "Super Net",
-        Kinsta => "Lost Maiamai",
-        BadgeBee => "Bee Badge",
-        HintGlasses => "Hint Glasses",
-        LiverBlue => "Monster Tail",
-        LiverPurple => "Monster Guts",
-        LiverYellow => "Monster Horn",
-        ClothesBlue | ClothesRed => "Armor Upgrade",
-        HyruleShield => "Hylian Shield",
-        OreYellow => "Master Ore",
-        OreGreen => "Master Ore",
-        OreBlue => "Master Ore",
-        GanbariPowerUp => "Stamina Scroll",
-        Pouch => "Pouch",
-        DashBoots => "Pegasus Boots",
-        OreRed => "Master Ore",
-        MessageBottle => "Letter in a Bottle",
-        MilkMatured => "Premium Milk",
-        SpecialMove => "Great Spin",
-        GanbariTubo => "Energy Potion",
-        RupeePurple => "Purple Rupee",
-        ItemBowLight => "Bow of Light",
-        Heart => "Heart",
-
-        Empty => "Nothing",
-
-        PendantPower => "Pendant of Power",
-        PendantWisdom => "Pendant of Wisdom",
-        ZeldaAmulet | PendantCourage => "Pendant of Courage Upgrade",
-
-        SageGulley => "Sage Gulley",
-        SageOren => "Sage Oren",
-        SageSeres => "Sage Seres",
-        SageOsfala => "Sage Osfala",
-        SageImpa => "Sage Impa",
-        SageIrene => "Sage Irene",
-        SageRosso => "Sage Rosso",
-
-        TriforceCourage => "Triforce of Courage",
-
-        ItemPotShopRed => "Red Potion",
-        ItemPotShopBlue => "Blue Potion",
-        ItemPotShopPurple => "Purple Potion",
-        ItemPotShopYellow => "Yellow Potion",
-
-        EscapeFruit => "Scoot Fruit",
-        StopFruit => "Foul Fruit",
-        Bee => "Bee",
-        GoldenBeeForSale => "Golden Bee",
-        Fairy => "Fairy",
-        Milk => "Milk",
-
-        ItemRentalIceRod => "Rented Ice Rod",
-        ItemRentalSandRod => "Rented Sand Rod",
-        ItemRentalTornadeRod => "Rented Tornado Rod",
-        ItemRentalBomb => "Rented Bomb Rod",
-        ItemRentalFireRod => "Rented Fire Rod",
-        ItemRentalHookShot => "Rented Hookshot",
-        ItemRentalBoomerang => "Rented Boomerang",
-        ItemRentalHammer => "Rented Hammer",
-        ItemRentalBow => "Rented Bow",
-        ItemRentalShield => "Rented Shield",
-        ItemRentalSandRodFirst => "Rented Sand Rod (Osfala)",
-        PowerfulGlove => "Titan's Mitt",
-        GoldenBee => "Golden Bee",
-        PackageSword => "Captain's Sword",
-    }
-}
-
-trait ItemExt {
-    fn normalize(self) -> Self;
-    fn goes_in_csmc_large_chest(&self) -> bool;
-    fn msbf_key(self) -> Result<Option<&'static str>>;
-
-    // fn is_dungeon(&self) -> bool;
-    // fn is_sword(&self) -> bool;
-    // fn is_super(&self) -> bool;
-    // fn is_ore(&self) -> bool;
-}
-
-impl ItemExt for Item {
-    fn normalize(self) -> Self {
-        match self {
-            PackageSword | ItemSwordLv1 | ItemSwordLv3 | ItemSwordLv4 => ItemSwordLv2,
-            ItemRentalIceRod => ItemIceRod,
-            ItemRentalSandRod => ItemSandRod,
-            ItemRentalTornadeRod => ItemTornadeRod,
-            ItemRentalBomb => ItemBomb,
-            ItemRentalFireRod => ItemFireRod,
-            ItemRentalHookShot => ItemHookShot,
-            ItemRentalBoomerang => ItemBoomerang,
-            ItemRentalHammer => ItemHammer,
-            ItemRentalBow => ItemBow,
-            PowerfulGlove => PowerGlove,
-            ClothesRed => ClothesBlue,
-            // RingRental => RingHekiga,
-            ItemKandelaarLv2 => ItemKandelaar,
-            ItemInsectNetLv2 => ItemInsectNet,
-            item => item,
-        }
-    }
-
-    fn goes_in_csmc_large_chest(&self) -> bool {
-        matches!(
-            self,
-            // Empty |
-            KeySmall | KeyBoss |
-            // Compass |
-            // HeartContainer | HeartPiece |
-            // RupeeR | RupeeG | RupeeB | RupeeGold | RupeeSilver | RupeePurple |
-            ItemIceRod | ItemRentalIceRod | ItemIceRodLv2 |
-            ItemSandRod | ItemRentalSandRod | ItemSandRodLv2 | ItemRentalSandRodFirst |
-            ItemTornadeRod | ItemRentalTornadeRod | ItemTornadeRodLv2 |
-            ItemBomb | ItemRentalBomb | ItemBombLv2 |
-            ItemFireRod | ItemRentalFireRod | ItemFireRodLv2 |
-            ItemHookShot | ItemRentalHookShot | ItemHookShotLv2 |
-            ItemBoomerang | ItemRentalBoomerang | ItemBoomerangLv2 |
-            ItemHammer | ItemRentalHammer | ItemHammerLv2 |
-            ItemBow | ItemRentalBow | ItemBowLv2 |
-            ItemShield | ItemRentalShield | HyruleShield |
-            ItemBottle |
-            // ItemPotShopRed | ItemPotShopBlue | ItemPotShopPurple | ItemPotShopYellow | Milk |
-            ItemStoneBeauty |
-            PendantPower | PendantWisdom | PendantCourage |
-            ZeldaAmulet |
-            ItemKandelaar | ItemKandelaarLv2 |
-            ItemSwordLv1 | ItemSwordLv2 | ItemSwordLv3 | ItemSwordLv4 | PackageSword |
-            ItemMizukaki |
-            RingRental | RingHekiga |
-            ItemBell |
-            PowerGlove | PowerfulGlove |
-            ItemInsectNet | ItemInsectNetLv2 |
-            // Kinsta |
-            BadgeBee |
-            GoldenBee |
-            // Bee | Fairy | GoldenBeeForSale |
-            HintGlasses |
-            EscapeFruit |
-            StopFruit |
-            // LiverBlue | LiverPurple | LiverYellow |
-            ClothesBlue | ClothesRed |
-            OreYellow | OreGreen | OreBlue | OreRed |
-            GanbariPowerUp |
-            // GanbariTubo |
-            Pouch |
-            DashBoots |
-            MessageBottle | MilkMatured |
-            SpecialMove |
-            ItemBowLight |
-            // TriforceCourage |
-            // Heart |
-            SageGulley | SageOren | SageSeres | SageOsfala | SageImpa | SageIrene | SageRosso
-        )
-    }
-
-    fn msbf_key(self) -> Result<Option<&'static str>> {
-        match self {
-            SageGulley => Ok(Some(MsbfKey::Dark)),
-            SageOren => Ok(Some(MsbfKey::Water)),
-            SageSeres => Ok(Some(MsbfKey::Dokuro)),
-            SageOsfala => Ok(Some(MsbfKey::Hagure)),
-            SageIrene => Ok(Some(MsbfKey::Sand)),
-            SageRosso => Ok(Some(MsbfKey::Ice)),
-            SageImpa => Ok(None), // Impa special
-            PendantPower | PendantWisdom | PendantCourage | ZeldaAmulet => Ok(None),
-            _ => Err(Error::new("")),
-        }
+impl From<modd::Error> for Error {
+    fn from(err: modd::Error) -> Self {
+        Self::Message(err.to_string())
     }
 }
