+++ rom/src/lib.rs
@@ -245,21 +245,3 @@ impl Rom {
 }
 
 const US_ID: u64 = 0x00040000000EC300;
-
-#[macro_export]
-macro_rules! string_constants {
-    (
-        $(#[$attr:meta])*
-        $type:ident {
-            $($variant:ident,)+
-        }
-    ) => {
-        $(#[$attr])*
-        pub struct $type;
-
-        $(#[$attr])*
-        impl $type {
-            $(pub const $variant: &'static str = stringify!($variant);)+
-        }
-    }
-}
